requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
